requirements.txt
# Zillow Scraper Dependencies
streamlit>=1.32.0
streamlit-echarts>=0.4.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.1.0

# Development
pytest>=7.4.0
black>=23.12.0
ruff>=0.1.0
